+++ doc/RELEASE-NOTES.md
@@ -2,6 +2,9 @@
 
 ## 0.2.10
 
+* Fix bug in Resource#cut which was not actually cutting
+  resource
+* Fix NaivePlanner was not taking disk resources into account (#165)
 * Support standard format of Mesos master discovery address
   e.g. `zk://zk1:2181,zk2:2181,zk3:2181/mesos` - with this
   now Retz survives Mesos master change and restart (#140, #141)